// src/common/exceptions/entities/enums/tag-exception.enum.ts
export enum TagExceptionMessages {
  AddItemTagException = 'Error adding item tag: ',
  DeleteItemTagException = 'Error deleting item tag: ',
  AddTagException = 'Error adding tag: ',
  DeleteTagException = 'Error deleting tag: ',
  FetchTagsException = 'Error getting tags: ',
  SetSelectedTagsException = 'Error setting selected tags: ',
  AddTagToSelectedException = 'Error adding tag to selected tags: ',
  RemoveTagFromSelectedException = 'Error removing tag from selected tags: ',
}

export enum TagExceptionNames {
  AddItemTagException = 'AddItemTagException',
  DeleteItemTagException = 'DeleteItemTagException',
  AddTagException = 'AddTagException',
  DeleteTagException = 'DeleteTagException',
  FetchTagsException = 'FetchTagsException',
  SetSelectedTagsException = 'SetSelectedTagsException',
  AddTagToSelectedException = 'AddTagToSelectedException',
  RemoveTagFromSelectedException = 'RemoveTagFromSelectedException',
}